fix: UnitBase quantity from float to int

+++ Meal Service/src/models/meal.py
@@ -1,5 +1,5 @@
 from datetime import date
-from sqlalchemy import Integer, Float, String, Boolean, Date, JSON, ForeignKey, Enum, UniqueConstraint
+from sqlalchemy import Integer, String, Boolean, Date, JSON, ForeignKey, Enum, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from database import Base
 from enums.meal_type import MealType
@@ -49,7 +49,7 @@ class StorableUnitList(Base):
     unit_id: Mapped[int] = mapped_column(Integer, primary_key=True)
     meal_id: Mapped[int] = mapped_column(ForeignKey("meals.meal_id"), primary_key=True)
     unit_name: Mapped[str] = mapped_column(String, nullable=False)
-    quantity: Mapped[float] = mapped_column(Float, nullable=False)
+    quantity: Mapped[int] = mapped_column(Integer, nullable=False)
 
     meal: Mapped["Meal"] = relationship(
         back_populates="storable_unit_list",
